Ability to enable tracing on a flow using a rule
Patch adds the ability to enable tracing on a flow by adding a
rule. This patch includes the updates to the rest api, the cli client
and midolman itself. 'enabling' trace for a flow doesn't do anything
yet, except tagging the packet context.

When a flow matches the condition of the trace rule, an exception it
thrown back to the dedupe actor to tell it to retry the packet context
with tracing enabled.

It is in the dedupe actor that we can add code to enable the actual
tracing once we have decided what the trace output will look like and
where it will go.

Change-Id: Ib679ef9669530827dc0d2e0641e187d1abfcbdb9